% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stashapi_functions.R
\name{addTempDLNAIP}
\alias{addTempDLNAIP}
\title{Call GraphQL operation: addTempDLNAIP}
\usage{
addTempDLNAIP(input = NA, ...)
}
\arguments{
\item{input}{See the Playground for further details.}

\item{...}{additional parameters like .field to retrieve only a certain field from the response}
}
\value{
processed API response
}
\description{
Enables an IP address for DLNA for an optional duration
}
