% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stashapi_functions.R
\name{allGalleries}
\alias{allGalleries}
\title{Call GraphQL operation: allGalleries}
\usage{
allGalleries(...)
}
\arguments{
\item{...}{Additional options such as `.field`, `.response`, and `.progress_bar`.}
}
\value{
The processed API response.
}
\description{
Executes the GraphQL operation `allGalleries`.
}
\details{
**Deprecated:** Use findGalleries instead
}
