% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stashapi_functions.R
\name{allMovies}
\alias{allMovies}
\title{Call GraphQL operation: allMovies}
\usage{
allMovies(...)
}
\arguments{
\item{...}{Additional options such as `.field`, `.response`, and `.progress_bar`.}
}
\value{
The processed API response.
}
\description{
Executes the GraphQL operation `allMovies`.
}
\details{
**Deprecated:** Use findGroups instead
}
