% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stashapi_functions.R
\name{allSceneMarkers}
\alias{allSceneMarkers}
\title{Call GraphQL operation: allSceneMarkers}
\usage{
allSceneMarkers(...)
}
\arguments{
\item{...}{Additional options such as `.field`, `.response`, and `.progress_bar`.}
}
\value{
The processed API response.
}
\description{
Executes the GraphQL operation `allSceneMarkers`.
}
\details{
**Deprecated:** Use findSceneMarkers instead
}
