% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stashapi_functions.R
\name{allScenes}
\alias{allScenes}
\title{Call GraphQL operation: allScenes}
\usage{
allScenes(...)
}
\arguments{
\item{...}{Additional options such as `.field`, `.response`, and `.progress_bar`.}
}
\value{
The processed API response.
}
\description{
Executes the GraphQL operation `allScenes`.
}
\details{
**Deprecated:** Use findScenes instead
}
