% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stashapi_functions.R
\name{allStudios}
\alias{allStudios}
\title{Call GraphQL operation: allStudios}
\usage{
allStudios(...)
}
\arguments{
\item{...}{Additional options such as `.field`, `.response`, and `.progress_bar`.}
}
\value{
The processed API response.
}
\description{
Executes the GraphQL operation `allStudios`.
}
\details{
**Deprecated:** Use findStudios instead
}
