% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stashapi_functions.R
\name{allTags}
\alias{allTags}
\title{Call GraphQL operation: allTags}
\usage{
allTags(...)
}
\arguments{
\item{...}{Additional options such as `.field`, `.response`, and `.progress_bar`.}
}
\value{
The processed API response.
}
\description{
Executes the GraphQL operation `allTags`.
}
\details{
**Deprecated:** Use findTags instead
}
