% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stashapi_functions.R
\name{anonymiseDatabase}
\alias{anonymiseDatabase}
\title{Call GraphQL operation: anonymiseDatabase}
\usage{
anonymiseDatabase(input = NA, ...)
}
\arguments{
\item{input}{See the Stash Playground for details.}

\item{...}{Additional options such as `.field`, `.response`, and `.progress_bar`.}
}
\value{
The processed API response.
}
\description{
Anonymise the database in a separate file. Optionally returns a link to download the database file
}
