% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stashapi_functions.R
\name{availablePackages}
\alias{availablePackages}
\title{Call GraphQL operation: availablePackages}
\usage{
availablePackages(type = NA, source = list(), ...)
}
\arguments{
\item{type}{See the Stash Playground for details.}

\item{source}{See the Stash Playground for details.}

\item{...}{Additional options such as `.field`, `.response`, and `.progress_bar`.}
}
\value{
The processed API response.
}
\description{
List available packages
}
