% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stashapi_functions.R
\name{backupDatabase}
\alias{backupDatabase}
\title{Call GraphQL operation: backupDatabase}
\usage{
backupDatabase(input = NA, ...)
}
\arguments{
\item{input}{See the Playground for further details.}

\item{...}{additional parameters like .field to retrieve only a certain field from the response}
}
\value{
processed API response
}
\description{
Backup the database. Optionally returns a link to download the database file
}
