% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stashapi_functions.R
\name{configuration}
\alias{configuration}
\title{Call GraphQL operation: configuration}
\usage{
configuration(...)
}
\arguments{
\item{...}{Additional options such as `.field`, `.response`, and `.progress_bar`.}
}
\value{
The processed API response.
}
\description{
Returns the current, complete configuration
}
