% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stashapi_functions.R
\name{configurePlugin}
\alias{configurePlugin}
\title{Call GraphQL operation: configurePlugin}
\usage{
configurePlugin(pluginid = list(), input = list(), ...)
}
\arguments{
\item{pluginid}{See the Playground for further details.}

\item{input}{See the Playground for further details.}

\item{...}{additional parameters like .field to retrieve only a certain field from the response}
}
\value{
processed API response
}
\description{
overwrites the entire plugin configuration for the given plugin
}
