% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stashapi_functions.R
\name{configureUI}
\alias{configureUI}
\title{Call GraphQL operation: configureUI}
\usage{
configureUI(input = NA, partial = NA, ...)
}
\arguments{
\item{input}{See the Stash Playground for details.}

\item{partial}{See the Stash Playground for details.}

\item{...}{Additional options such as `.field`, `.response`, and `.progress_bar`.}
}
\value{
The processed API response.
}
\description{
overwrites the UI configuration
if input is provided, then the entire UI configuration is replaced
if partial is provided, then the partial UI configuration is merged into the existing UI configuration
}
