% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stashapi_functions.R
\name{configureUISetting}
\alias{configureUISetting}
\title{Call GraphQL operation: configureUISetting}
\usage{
configureUISetting(key = list(), value = NA, ...)
}
\arguments{
\item{key}{See the Stash Playground for details.}

\item{value}{See the Stash Playground for details.}

\item{...}{Additional options such as `.field`, `.response`, and `.progress_bar`.}
}
\value{
The processed API response.
}
\description{
sets a single UI key value
key is a dot separated path to the value
}
