% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stashapi_functions.R
\name{destroyFiles}
\alias{destroyFiles}
\title{Call GraphQL operation: destroyFiles}
\usage{
destroyFiles(ids = list(), ...)
}
\arguments{
\item{ids}{See the Playground for further details.}

\item{...}{additional parameters like .field to retrieve only a certain field from the response}
}
\value{
processed API response
}
\description{
Deletes file entries from the database without deleting the files from the filesystem
}
