% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stashapi_functions.R
\name{destroySavedFilter}
\alias{destroySavedFilter}
\title{Call GraphQL operation: destroySavedFilter}
\usage{
destroySavedFilter(input = NA, ...)
}
\arguments{
\item{input}{See the Stash Playground for details.}

\item{...}{Additional options such as `.field`, `.response`, and `.progress_bar`.}
}
\value{
The processed API response.
}
\description{
Executes the GraphQL operation `destroySavedFilter`.
}
