% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stashapi_functions.R
\name{directory}
\alias{directory}
\title{Call GraphQL operation: directory}
\usage{
directory(path = NA, locale = NA, ...)
}
\arguments{
\item{path}{The directory path to list}

\item{locale}{Desired collation locale. Determines the order of the directory result. eg. 'en-US', 'pt-BR', ...}

\item{...}{Additional options such as `.field`, `.response`, and `.progress_bar`.}
}
\value{
The processed API response.
}
\description{
Returns an array of paths for the given path
}
