% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stashapi_functions.R
\name{disableDLNA}
\alias{disableDLNA}
\title{Call GraphQL operation: disableDLNA}
\usage{
disableDLNA(input = NA, ...)
}
\arguments{
\item{input}{See the Stash Playground for details.}

\item{...}{Additional options such as `.field`, `.response`, and `.progress_bar`.}
}
\value{
The processed API response.
}
\description{
Disables DLNA for an optional duration. Has no effect if DLNA is disabled by default
}
