% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stashapi_functions.R
\name{dlnaStatus}
\alias{dlnaStatus}
\title{Call GraphQL operation: dlnaStatus}
\usage{
dlnaStatus(...)
}
\arguments{
\item{...}{additional parameters like .field to retrieve only a certain field from the response}
}
\value{
processed API response
}
\description{
Call GraphQL operation: dlnaStatus
}
