% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stashapi_functions.R
\name{downloadFFMpeg}
\alias{downloadFFMpeg}
\title{Call GraphQL operation: downloadFFMpeg}
\usage{
downloadFFMpeg(...)
}
\arguments{
\item{...}{additional parameters like .field to retrieve only a certain field from the response}
}
\value{
processed API response
}
\description{
Downloads and installs ffmpeg and ffprobe binaries into the configuration directory. Returns the job ID.
}
