% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stashapi_functions.R
\name{enableDLNA}
\alias{enableDLNA}
\title{Call GraphQL operation: enableDLNA}
\usage{
enableDLNA(input = NA, ...)
}
\arguments{
\item{input}{See the Stash Playground for details.}

\item{...}{Additional options such as `.field`, `.response`, and `.progress_bar`.}
}
\value{
The processed API response.
}
\description{
Enables DLNA for an optional duration. Has no effect if DLNA is enabled by default
}
