% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_helpers.R
\name{equals}
\alias{equals}
\title{Build an EQUALS criterion.}
\usage{equals(value)}
\arguments{\item{value}{Criterion value.}}
\value{A named criterion list.}
\description{Builds a GraphQL EQUALS criterion.}
\examples{equals("4k")}
