% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_helpers.R
\name{excludes}
\alias{excludes}
\title{Build an EXCLUDES criterion.}
\usage{
excludes(value, depth = NULL)
}
\arguments{
\item{value}{IDs or values to exclude.}

\item{depth}{Optional hierarchy depth. `-1` means unlimited depth.}
}
\description{
Build an EXCLUDES criterion.
}
\examples{excludes(182, depth = -1)}
