% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stashapi_functions.R
\name{findDefaultFilter}
\alias{findDefaultFilter}
\title{Call GraphQL operation: findDefaultFilter}
\usage{
findDefaultFilter(mode = NA, ...)
}
\arguments{
\item{mode}{See the Stash Playground for details.}

\item{...}{Additional options such as `.field`, `.response`, and `.progress_bar`.}
}
\value{
The processed API response.
}
\description{
Executes the GraphQL operation `findDefaultFilter`.
}
\details{
**Deprecated:** default filter now stored in UI config
}
