% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stashapi_functions.R
\name{findFiles}
\alias{findFiles}
\title{Call GraphQL operation: findFiles}
\usage{
findFiles(filefilter = NA, filter = NA, ids = list(), ...)
}
\arguments{
\item{filefilter}{See the Stash Playground for details.}

\item{filter}{See the Stash Playground for details.}

\item{ids}{See the Stash Playground for details.}

\item{...}{Additional options such as `.field`, `.response`, and `.progress_bar`.}
}
\value{
The processed API response.
}
\description{
Queries for Files
}
