% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stashapi_functions.R
\name{findFolder}
\alias{findFolder}
\title{Call GraphQL operation: findFolder}
\usage{
findFolder(id = NA, path = NA, ...)
}
\arguments{
\item{id}{See the Stash Playground for details.}

\item{path}{See the Stash Playground for details.}

\item{...}{Additional options such as `.field`, `.response`, and `.progress_bar`.}
}
\value{
The processed API response.
}
\description{
Find a file by its id or path
}
