% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stashapi_functions.R
\name{findFolders}
\alias{findFolders}
\title{Call GraphQL operation: findFolders}
\usage{
findFolders(folderfilter = NA, filter = NA, ids = list(), ...)
}
\arguments{
\item{folderfilter}{See the Stash Playground for details.}

\item{filter}{See the Stash Playground for details.}

\item{ids}{See the Stash Playground for details.}

\item{...}{Additional options such as `.field`, `.response`, and `.progress_bar`.}
}
\value{
The processed API response.
}
\description{
Queries for Files
}
