% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stashapi_functions.R
\name{findImage}
\alias{findImage}
\title{Call GraphQL operation: findImage}
\usage{
findImage(id = NA, checksum = NA, ...)
}
\arguments{
\item{id}{See the Stash Playground for details.}

\item{checksum}{See the Stash Playground for details.}

\item{...}{Additional options such as `.field`, `.response`, and `.progress_bar`.}
}
\value{
The processed API response.
}
\description{
Executes the GraphQL operation `findImage`.
}
