% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stashapi_functions.R
\name{findImages}
\alias{findImages}
\title{Call GraphQL operation: findImages}
\usage{
findImages(imagefilter = NA, imageids = list(), ids = list(), filter = NA, ...)
}
\arguments{
\item{imagefilter}{See the Playground for further details.}

\item{imageids}{See the Playground for further details.}

\item{ids}{See the Playground for further details.}

\item{filter}{See the Playground for further details.}

\item{...}{additional parameters including \code{.field} for field selection, \code{.response} for response mode, and \code{.progress_bar} to show an opt-in progress bar while collecting paginated results.}
}
\value{
processed API response
}
\description{
A function which queries Scene objects
}
