% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stashapi_functions.R
\name{findMovie}
\alias{findMovie}
\title{Call GraphQL operation: findMovie}
\usage{
findMovie(id = list(), ...)
}
\arguments{
\item{id}{See the Stash Playground for details.}

\item{...}{Additional options such as `.field`, `.response`, and `.progress_bar`.}
}
\value{
The processed API response.
}
\description{
Find a movie by ID
}
\details{
**Deprecated:** Use findGroup instead
}
