% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stashapi_functions.R
\name{findMovies}
\alias{findMovies}
\title{Call GraphQL operation: findMovies}
\usage{
findMovies(moviefilter = NA, filter = NA, ids = list(), ...)
}
\arguments{
\item{moviefilter}{See the Stash Playground for details.}

\item{filter}{See the Stash Playground for details.}

\item{ids}{See the Stash Playground for details.}

\item{...}{Additional options such as `.field`, `.response`, and `.progress_bar`.}
}
\value{
The processed API response.
}
\description{
A function which queries Movie objects
}
\details{
**Deprecated:** Use findGroups instead
}
