% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stashapi_functions.R
\name{findSavedFilter}
\alias{findSavedFilter}
\title{Call GraphQL operation: findSavedFilter}
\usage{
findSavedFilter(id = list(), ...)
}
\arguments{
\item{id}{See the Stash Playground for details.}

\item{...}{Additional options such as `.field`, `.response`, and `.progress_bar`.}
}
\value{
The processed API response.
}
\description{
Executes the GraphQL operation `findSavedFilter`.
}
