% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stashapi_functions.R
\name{findScene}
\alias{findScene}
\title{Call GraphQL operation: findScene}
\usage{
findScene(id = NA, checksum = NA, ...)
}
\arguments{
\item{id}{See the Stash Playground for details.}

\item{checksum}{See the Stash Playground for details.}

\item{...}{Additional options such as `.field`, `.response`, and `.progress_bar`.}
}
\value{
The processed API response.
}
\description{
Find a scene by ID or Checksum
}
