% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stashapi_functions.R
\name{findScenes}
\alias{findScenes}
\title{Call GraphQL operation: findScenes}
\usage{
findScenes(scenefilter = NA, sceneids = list(), ids = list(), filter = NA, ...)
}
\arguments{
\item{scenefilter}{See the Stash Playground for details.}

\item{sceneids}{See the Stash Playground for details.}

\item{ids}{See the Stash Playground for details.}

\item{filter}{See the Stash Playground for details.}

\item{...}{Additional options such as `.field`, `.response`, and `.progress_bar`.}
}
\value{
The processed API response.
}
\description{
A function which queries Scene objects
}
