% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stashapi_functions.R
\name{findScenesByPathRegex}
\alias{findScenesByPathRegex}
\title{Call GraphQL operation: findScenesByPathRegex}
\usage{
findScenesByPathRegex(filter = NA, ...)
}
\arguments{
\item{filter}{See the Playground for further details.}

\item{...}{additional parameters like .field to retrieve only a certain field from the response}
}
\value{
processed API response
}
\description{
Call GraphQL operation: findScenesByPathRegex
}
