% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stashapi_functions.R
\name{findStudios}
\alias{findStudios}
\title{Call GraphQL operation: findStudios}
\usage{
findStudios(studiofilter = NA, filter = NA, ids = list(), ...)
}
\arguments{
\item{studiofilter}{See the Stash Playground for details.}

\item{filter}{See the Stash Playground for details.}

\item{ids}{See the Stash Playground for details.}

\item{...}{Additional options such as `.field`, `.response`, and `.progress_bar`.}
}
\value{
The processed API response.
}
\description{
A function which queries Studio objects
}
