% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stashapi_functions.R
\name{findTags}
\alias{findTags}
\title{Call GraphQL operation: findTags}
\usage{
findTags(tagfilter = NA, filter = NA, ids = list(), ...)
}
\arguments{
\item{tagfilter}{See the Playground for further details.}

\item{filter}{See the Playground for further details.}

\item{ids}{See the Playground for further details.}

\item{...}{additional parameters like .field to retrieve only a certain field from the response}
}
\value{
processed API response
}
\description{
Call GraphQL operation: findTags
}
