% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_helpers.R
\name{gallery_filter}
\alias{gallery_filter}
\title{Build a gallery filter.}
\usage{gallery_filter(..., .strict = TRUE)}
\arguments{\item{...}{Filter fields from the Stash `GalleryFilterType` input.}\item{.strict}{Whether unknown fields should cause an error.}}
\value{A named list suitable for `findGalleries(galleryfilter = ...)`.}
