% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stashapi_functions.R
\name{generateAPIKey}
\alias{generateAPIKey}
\title{Call GraphQL operation: generateAPIKey}
\usage{
generateAPIKey(input = NA, ...)
}
\arguments{
\item{input}{See the Stash Playground for details.}

\item{...}{Additional options such as `.field`, `.response`, and `.progress_bar`.}
}
\value{
The processed API response.
}
\description{
Generate and set (or clear) API key
}
