% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stashapi_functions.R
\name{imageDecrementO}
\alias{imageDecrementO}
\title{Call GraphQL operation: imageDecrementO}
\usage{
imageDecrementO(id = list(), ...)
}
\arguments{
\item{id}{See the Stash Playground for details.}

\item{...}{Additional options such as `.field`, `.response`, and `.progress_bar`.}
}
\value{
The processed API response.
}
\description{
Decrements the o-counter for an image. Returns the new value
}
