% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stashapi_functions.R
\name{imageIncrementO}
\alias{imageIncrementO}
\title{Call GraphQL operation: imageIncrementO}
\usage{
imageIncrementO(id = list(), ...)
}
\arguments{
\item{id}{See the Stash Playground for details.}

\item{...}{Additional options such as `.field`, `.response`, and `.progress_bar`.}
}
\value{
The processed API response.
}
\description{
Increments the o-counter for an image. Returns the new value
}
