% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stashapi_functions.R
\name{imageResetO}
\alias{imageResetO}
\title{Call GraphQL operation: imageResetO}
\usage{
imageResetO(id = list(), ...)
}
\arguments{
\item{id}{See the Playground for further details.}

\item{...}{additional parameters like .field to retrieve only a certain field from the response}
}
\value{
processed API response
}
\description{
Resets the o-counter for a image to 0. Returns the new value
}
