% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stashapi_functions.R
\name{imagesDestroy}
\alias{imagesDestroy}
\title{Call GraphQL operation: imagesDestroy}
\usage{
imagesDestroy(input = NA, ...)
}
\arguments{
\item{input}{See the Stash Playground for details.}

\item{...}{Additional options such as `.field`, `.response`, and `.progress_bar`.}
}
\value{
The processed API response.
}
\description{
Executes the GraphQL operation `imagesDestroy`.
}
