% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stashapi_functions.R
\name{importObjects}
\alias{importObjects}
\title{Call GraphQL operation: importObjects}
\usage{
importObjects(input = NA, ...)
}
\arguments{
\item{input}{See the Playground for further details.}

\item{...}{additional parameters like .field to retrieve only a certain field from the response}
}
\value{
processed API response
}
\description{
Performs an incremental import. Returns the job ID
}
