% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stashapi_functions.R
\name{installedPackages}
\alias{installedPackages}
\title{Call GraphQL operation: installedPackages}
\usage{
installedPackages(type = NA, ...)
}
\arguments{
\item{type}{See the Stash Playground for details.}

\item{...}{Additional options such as `.field`, `.response`, and `.progress_bar`.}
}
\value{
The processed API response.
}
\description{
List installed packages
}
