% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection.R
\name{is_stash_connected}
\alias{is_stash_connected}
\title{Check if the active Stash connection is valid}
\usage{
is_stash_connected()
}
\value{
A logical value: `TRUE` for a valid active connection and `FALSE` otherwise.
}
\description{
Validates whether the active connection to the Stash GraphQL endpoint is functional.
}
\examples{
\dontrun{
is_stash_connected()
}
}
\keyword{connection}
\keyword{validation}