% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stashapi_functions.R
\name{listScrapers}
\alias{listScrapers}
\title{Call GraphQL operation: listScrapers}
\usage{
listScrapers(types = NA, ...)
}
\arguments{
\item{types}{See the Stash Playground for details.}

\item{...}{Additional options such as `.field`, `.response`, and `.progress_bar`.}
}
\value{
The processed API response.
}
\description{
List available scrapers
}
