% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stashapi_functions.R
\name{logs}
\alias{logs}
\title{Call GraphQL operation: logs}
\usage{
logs(...)
}
\arguments{
\item{...}{Additional options such as `.field`, `.response`, and `.progress_bar`.}
}
\value{
The processed API response.
}
\description{
Executes the GraphQL operation `logs`.
}
