% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stashapi_functions.R
\name{markerStrings}
\alias{markerStrings}
\title{Call GraphQL operation: markerStrings}
\usage{
markerStrings(q = NA, sort = NA, ...)
}
\arguments{
\item{q}{See the Stash Playground for details.}

\item{sort}{See the Stash Playground for details.}

\item{...}{Additional options such as `.field`, `.response`, and `.progress_bar`.}
}
\value{
The processed API response.
}
\description{
Get marker strings
}
