% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stashapi_functions.R
\name{markerWall}
\alias{markerWall}
\title{Call GraphQL operation: markerWall}
\usage{
markerWall(q = NA, ...)
}
\arguments{
\item{q}{See the Stash Playground for details.}

\item{...}{Additional options such as `.field`, `.response`, and `.progress_bar`.}
}
\value{
The processed API response.
}
\description{
Retrieve random scene markers for the wall
}
