% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stashapi_functions.R
\name{metadataAutoTag}
\alias{metadataAutoTag}
\title{Call GraphQL operation: metadataAutoTag}
\usage{
metadataAutoTag(input = NA, ...)
}
\arguments{
\item{input}{See the Playground for further details.}

\item{...}{additional parameters like .field to retrieve only a certain field from the response}
}
\value{
processed API response
}
\description{
Start auto-tagging. Returns the job ID
}
