% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stashapi_functions.R
\name{metadataClean}
\alias{metadataClean}
\title{Call GraphQL operation: metadataClean}
\usage{
metadataClean(input = NA, ...)
}
\arguments{
\item{input}{See the Stash Playground for details.}

\item{...}{Additional options such as `.field`, `.response`, and `.progress_bar`.}
}
\value{
The processed API response.
}
\description{
Clean metadata. Returns the job ID
}
