% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stashapi_functions.R
\name{metadataCleanGenerated}
\alias{metadataCleanGenerated}
\title{Call GraphQL operation: metadataCleanGenerated}
\usage{
metadataCleanGenerated(input = NA, ...)
}
\arguments{
\item{input}{See the Stash Playground for details.}

\item{...}{Additional options such as `.field`, `.response`, and `.progress_bar`.}
}
\value{
The processed API response.
}
\description{
Clean generated files. Returns the job ID
}
