% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stashapi_functions.R
\name{metadataExport}
\alias{metadataExport}
\title{Call GraphQL operation: metadataExport}
\usage{
metadataExport(...)
}
\arguments{
\item{...}{additional parameters like .field to retrieve only a certain field from the response}
}
\value{
processed API response
}
\description{
Start a full export. Outputs to the metadata directory. Returns the job ID
}
