% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stashapi_functions.R
\name{metadataIdentify}
\alias{metadataIdentify}
\title{Call GraphQL operation: metadataIdentify}
\usage{
metadataIdentify(input = NA, ...)
}
\arguments{
\item{input}{See the Stash Playground for details.}

\item{...}{Additional options such as `.field`, `.response`, and `.progress_bar`.}
}
\value{
The processed API response.
}
\description{
Identifies scenes using scrapers. Returns the job ID
}
