% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stashapi_functions.R
\name{metadataScan}
\alias{metadataScan}
\title{Call GraphQL operation: metadataScan}
\usage{
metadataScan(input = NA, ...)
}
\arguments{
\item{input}{See the Stash Playground for details.}

\item{...}{Additional options such as `.field`, `.response`, and `.progress_bar`.}
}
\value{
The processed API response.
}
\description{
Start a scan. Returns the job ID
}
