% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stashapi_functions.R
\name{migrate}
\alias{migrate}
\title{Call GraphQL operation: migrate}
\usage{
migrate(input = NA, ...)
}
\arguments{
\item{input}{See the Playground for further details.}

\item{...}{additional parameters like .field to retrieve only a certain field from the response}
}
\value{
processed API response
}
\description{
Migrates the schema to the required version. Returns the job ID
}
