% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stashapi_functions.R
\name{migrateBlobs}
\alias{migrateBlobs}
\title{Call GraphQL operation: migrateBlobs}
\usage{
migrateBlobs(input = NA, ...)
}
\arguments{
\item{input}{See the Stash Playground for details.}

\item{...}{Additional options such as `.field`, `.response`, and `.progress_bar`.}
}
\value{
The processed API response.
}
\description{
Migrates blobs from the old storage system to the current one
}
