% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stashapi_functions.R
\name{migrateSceneScreenshots}
\alias{migrateSceneScreenshots}
\title{Call GraphQL operation: migrateSceneScreenshots}
\usage{
migrateSceneScreenshots(input = NA, ...)
}
\arguments{
\item{input}{See the Stash Playground for details.}

\item{...}{Additional options such as `.field`, `.response`, and `.progress_bar`.}
}
\value{
The processed API response.
}
\description{
Migrates legacy scene screenshot files into the blob storage
}
