% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stashapi_functions.R
\name{movieCreate}
\alias{movieCreate}
\title{Call GraphQL operation: movieCreate}
\usage{
movieCreate(input = NA, ...)
}
\arguments{
\item{input}{See the Stash Playground for details.}

\item{...}{Additional options such as `.field`, `.response`, and `.progress_bar`.}
}
\value{
The processed API response.
}
\description{
Executes the GraphQL operation `movieCreate`.
}
\details{
**Deprecated:** Use groupCreate instead
}
