% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stashapi_functions.R
\name{parseSceneFilenames}
\alias{parseSceneFilenames}
\title{Call GraphQL operation: parseSceneFilenames}
\usage{
parseSceneFilenames(filter = NA, config = NA, ...)
}
\arguments{
\item{filter}{See the Stash Playground for details.}

\item{config}{See the Stash Playground for details.}

\item{...}{Additional options such as `.field`, `.response`, and `.progress_bar`.}
}
\value{
The processed API response.
}
\description{
Executes the GraphQL operation `parseSceneFilenames`.
}
