% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stashapi_functions.R
\name{performersDestroy}
\alias{performersDestroy}
\title{Call GraphQL operation: performersDestroy}
\usage{
performersDestroy(ids = list(), ...)
}
\arguments{
\item{ids}{See the Stash Playground for details.}

\item{...}{Additional options such as `.field`, `.response`, and `.progress_bar`.}
}
\value{
The processed API response.
}
\description{
Executes the GraphQL operation `performersDestroy`.
}
