% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stashapi_functions.R
\name{plugins}
\alias{plugins}
\title{Call GraphQL operation: plugins}
\usage{
plugins(...)
}
\arguments{
\item{...}{Additional options such as `.field`, `.response`, and `.progress_bar`.}
}
\value{
The processed API response.
}
\description{
List loaded plugins
}
