% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stashapi_functions.R
\name{pluginTasks}
\alias{pluginTasks}
\title{Call GraphQL operation: pluginTasks}
\usage{
pluginTasks(...)
}
\arguments{
\item{...}{Additional options such as `.field`, `.response`, and `.progress_bar`.}
}
\value{
The processed API response.
}
\description{
List available plugin operations
}
