% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stashapi_functions.R
\name{querySQL}
\alias{querySQL}
\title{Call GraphQL operation: querySQL}
\usage{
querySQL(sql = list(), args = NA, ...)
}
\arguments{
\item{sql}{See the Playground for further details.}

\item{args}{See the Playground for further details.}

\item{...}{additional parameters like .field to retrieve only a certain field from the response}
}
\value{
processed API response
}
\description{
DANGEROUS: Execute an arbitrary SQL statement that returns rows.
}
