% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stashapi_functions.R
\name{removeGalleryImages}
\alias{removeGalleryImages}
\title{Call GraphQL operation: removeGalleryImages}
\usage{
removeGalleryImages(input = NA, ...)
}
\arguments{
\item{input}{See the Stash Playground for details.}

\item{...}{Additional options such as `.field`, `.response`, and `.progress_bar`.}
}
\value{
The processed API response.
}
\description{
Executes the GraphQL operation `removeGalleryImages`.
}
