% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stashapi_functions.R
\name{removeTempDLNAIP}
\alias{removeTempDLNAIP}
\title{Call GraphQL operation: removeTempDLNAIP}
\usage{
removeTempDLNAIP(input = NA, ...)
}
\arguments{
\item{input}{See the Stash Playground for details.}

\item{...}{Additional options such as `.field`, `.response`, and `.progress_bar`.}
}
\value{
The processed API response.
}
\description{
Removes an IP address from the temporary DLNA whitelist
}
