% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stashapi_functions.R
\name{reorderSubGroups}
\alias{reorderSubGroups}
\title{Call GraphQL operation: reorderSubGroups}
\usage{
reorderSubGroups(input = NA, ...)
}
\arguments{
\item{input}{See the Stash Playground for details.}

\item{...}{Additional options such as `.field`, `.response`, and `.progress_bar`.}
}
\value{
The processed API response.
}
\description{
Reorder sub groups within a group. Returns true if successful.
}
