% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stashapi_functions.R
\name{revealFolderInFileManager}
\alias{revealFolderInFileManager}
\title{Call GraphQL operation: revealFolderInFileManager}
\usage{
revealFolderInFileManager(id = list(), ...)
}
\arguments{
\item{id}{See the Stash Playground for details.}

\item{...}{Additional options such as `.field`, `.response`, and `.progress_bar`.}
}
\value{
The processed API response.
}
\description{
Reveal the folder in the system file manager
}
