% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stashapi_functions.R
\name{runPluginOperation}
\alias{runPluginOperation}
\title{Call GraphQL operation: runPluginOperation}
\usage{
runPluginOperation(pluginid = list(), args = NA, ...)
}
\arguments{
\item{pluginid}{See the Playground for further details.}

\item{args}{See the Playground for further details.}

\item{...}{additional parameters like .field to retrieve only a certain field from the response}
}
\value{
processed API response
}
\description{
Runs a plugin operation. The operation is run immediately and does not use the job queue.
Returns a map of the result.
}
