% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stashapi_functions.R
\name{runPluginTask}
\alias{runPluginTask}
\title{Call GraphQL operation: runPluginTask}
\usage{
runPluginTask(
  pluginid = list(),
  taskname = NA,
  description = NA,
  args = NA,
  argsmap = NA,
  ...
)
}
\arguments{
\item{pluginid}{See the Playground for further details.}

\item{taskname}{if provided, then the default args will be applied}

\item{description}{displayed in the task queue}

\item{args}{See the Playground for further details.}

\item{argsmap}{See the Playground for further details.}

\item{...}{additional parameters like .field to retrieve only a certain field from the response}
}
\value{
processed API response
}
\description{
Run a plugin task.
If task_name is provided, then the task must exist in the plugin config and the tasks configuration
will be used to run the plugin.
If no task_name is provided, then the plugin will be executed with the arguments provided only.
Returns the job ID
}
