% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stashapi_functions.R
\name{sceneAddO}
\alias{sceneAddO}
\title{Call GraphQL operation: sceneAddO}
\usage{
sceneAddO(id = list(), times = list(), ...)
}
\arguments{
\item{id}{See the Playground for further details.}

\item{times}{See the Playground for further details.}

\item{...}{additional parameters like .field to retrieve only a certain field from the response}
}
\value{
processed API response
}
\description{
Increments the o-counter for a scene. Uses the current time if none provided.
}
