% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stashapi_functions.R
\name{sceneAddPlay}
\alias{sceneAddPlay}
\title{Call GraphQL operation: sceneAddPlay}
\usage{
sceneAddPlay(id = list(), times = list(), ...)
}
\arguments{
\item{id}{See the Stash Playground for details.}

\item{times}{See the Stash Playground for details.}

\item{...}{Additional options such as `.field`, `.response`, and `.progress_bar`.}
}
\value{
The processed API response.
}
\description{
Increments the play count for the scene. Uses the current time if none provided.
}
