% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stashapi_functions.R
\name{sceneDecrementO}
\alias{sceneDecrementO}
\title{Call GraphQL operation: sceneDecrementO}
\usage{
sceneDecrementO(id = list(), ...)
}
\arguments{
\item{id}{See the Stash Playground for details.}

\item{...}{Additional options such as `.field`, `.response`, and `.progress_bar`.}
}
\value{
The processed API response.
}
\description{
Decrements the o-counter for a scene. Returns the new value
}
\details{
**Deprecated:** Use sceneRemoveO instead
}
