% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stashapi_functions.R
\name{sceneDeleteO}
\alias{sceneDeleteO}
\title{Call GraphQL operation: sceneDeleteO}
\usage{
sceneDeleteO(id = list(), times = list(), ...)
}
\arguments{
\item{id}{See the Playground for further details.}

\item{times}{See the Playground for further details.}

\item{...}{additional parameters like .field to retrieve only a certain field from the response}
}
\value{
processed API response
}
\description{
Decrements the o-counter for a scene, removing the last recorded time if specific time not provided. Returns the new value
}
