% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stashapi_functions.R
\name{sceneGenerateScreenshot}
\alias{sceneGenerateScreenshot}
\title{Call GraphQL operation: sceneGenerateScreenshot}
\usage{
sceneGenerateScreenshot(id = list(), at = NA, ...)
}
\arguments{
\item{id}{See the Playground for further details.}

\item{at}{See the Playground for further details.}

\item{...}{additional parameters like .field to retrieve only a certain field from the response}
}
\value{
processed API response
}
\description{
Generates screenshot at specified time in seconds. Leave empty to generate default screenshot
}
