% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stashapi_functions.R
\name{sceneIncrementO}
\alias{sceneIncrementO}
\title{Call GraphQL operation: sceneIncrementO}
\usage{
sceneIncrementO(id = list(), ...)
}
\arguments{
\item{id}{See the Stash Playground for details.}

\item{...}{Additional options such as `.field`, `.response`, and `.progress_bar`.}
}
\value{
The processed API response.
}
\description{
Increments the o-counter for a scene. Returns the new value
}
\details{
**Deprecated:** Use sceneAddO instead
}
