% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stashapi_functions.R
\name{sceneIncrementPlayCount}
\alias{sceneIncrementPlayCount}
\title{Call GraphQL operation: sceneIncrementPlayCount}
\usage{
sceneIncrementPlayCount(id = list(), ...)
}
\arguments{
\item{id}{See the Stash Playground for details.}

\item{...}{Additional options such as `.field`, `.response`, and `.progress_bar`.}
}
\value{
The processed API response.
}
\description{
Increments the play count for the scene. Returns the new play count value.
}
\details{
**Deprecated:** Use sceneAddPlay instead
}
