% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stashapi_functions.R
\name{sceneMarkerDestroy}
\alias{sceneMarkerDestroy}
\title{Call GraphQL operation: sceneMarkerDestroy}
\usage{
sceneMarkerDestroy(id = list(), ...)
}
\arguments{
\item{id}{See the Playground for further details.}

\item{...}{additional parameters like .field to retrieve only a certain field from the response}
}
\value{
processed API response
}
\description{
Call GraphQL operation: sceneMarkerDestroy
}
