% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stashapi_functions.R
\name{sceneMarkersDestroy}
\alias{sceneMarkersDestroy}
\title{Call GraphQL operation: sceneMarkersDestroy}
\usage{
sceneMarkersDestroy(ids = list(), ...)
}
\arguments{
\item{ids}{See the Stash Playground for details.}

\item{...}{Additional options such as `.field`, `.response`, and `.progress_bar`.}
}
\value{
The processed API response.
}
\description{
Executes the GraphQL operation `sceneMarkersDestroy`.
}
