% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stashapi_functions.R
\name{sceneMarkerTags}
\alias{sceneMarkerTags}
\title{Call GraphQL operation: sceneMarkerTags}
\usage{
sceneMarkerTags(sceneid = list(), ...)
}
\arguments{
\item{sceneid}{See the Stash Playground for details.}

\item{...}{Additional options such as `.field`, `.response`, and `.progress_bar`.}
}
\value{
The processed API response.
}
\description{
Organize scene markers by tag for a given scene ID
}
