% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stashapi_functions.R
\name{sceneResetActivity}
\alias{sceneResetActivity}
\title{Call GraphQL operation: sceneResetActivity}
\usage{
sceneResetActivity(id = list(), resetresume = NA, resetduration = NA, ...)
}
\arguments{
\item{id}{See the Stash Playground for details.}

\item{resetresume}{See the Stash Playground for details.}

\item{resetduration}{See the Stash Playground for details.}

\item{...}{Additional options such as `.field`, `.response`, and `.progress_bar`.}
}
\value{
The processed API response.
}
\description{
Resets the resume time point and play duration
}
