% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stashapi_functions.R
\name{sceneResetO}
\alias{sceneResetO}
\title{Call GraphQL operation: sceneResetO}
\usage{
sceneResetO(id = list(), ...)
}
\arguments{
\item{id}{See the Stash Playground for details.}

\item{...}{Additional options such as `.field`, `.response`, and `.progress_bar`.}
}
\value{
The processed API response.
}
\description{
Resets the o-counter for a scene to 0. Returns the new value
}
