% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stashapi_functions.R
\name{sceneSaveActivity}
\alias{sceneSaveActivity}
\title{Call GraphQL operation: sceneSaveActivity}
\usage{
sceneSaveActivity(id = list(), resumetime = NA, playDuration = NA, ...)
}
\arguments{
\item{id}{See the Playground for further details.}

\item{resumetime}{See the Playground for further details.}

\item{playDuration}{See the Playground for further details.}

\item{...}{additional parameters like .field to retrieve only a certain field from the response}
}
\value{
processed API response
}
\description{
Sets the resume time point (if provided) and adds the provided duration to the scene's play duration
}
