% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stashapi_functions.R
\name{sceneStreams}
\alias{sceneStreams}
\title{Call GraphQL operation: sceneStreams}
\usage{
sceneStreams(id = NA, ...)
}
\arguments{
\item{id}{See the Stash Playground for details.}

\item{...}{Additional options such as `.field`, `.response`, and `.progress_bar`.}
}
\value{
The processed API response.
}
\description{
Return valid stream paths
}
