% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stashapi_functions.R
\name{sceneWall}
\alias{sceneWall}
\title{Call GraphQL operation: sceneWall}
\usage{
sceneWall(q = NA, ...)
}
\arguments{
\item{q}{See the Playground for further details.}

\item{...}{additional parameters like .field to retrieve only a certain field from the response}
}
\value{
processed API response
}
\description{
Retrieve random scenes for the wall
}
