% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stashapi_functions.R
\name{scrapeGroupURL}
\alias{scrapeGroupURL}
\title{Call GraphQL operation: scrapeGroupURL}
\usage{
scrapeGroupURL(url = list(), ...)
}
\arguments{
\item{url}{See the Playground for further details.}

\item{...}{additional parameters like .field to retrieve only a certain field from the response}
}
\value{
processed API response
}
\description{
Scrapes a complete group record based on a URL
}
