% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stashapi_functions.R
\name{scrapeImageURL}
\alias{scrapeImageURL}
\title{Call GraphQL operation: scrapeImageURL}
\usage{
scrapeImageURL(url = list(), ...)
}
\arguments{
\item{url}{See the Stash Playground for details.}

\item{...}{Additional options such as `.field`, `.response`, and `.progress_bar`.}
}
\value{
The processed API response.
}
\description{
Scrapes a complete image record based on a URL
}
