% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stashapi_functions.R
\name{scrapeMovieURL}
\alias{scrapeMovieURL}
\title{Call GraphQL operation: scrapeMovieURL}
\usage{
scrapeMovieURL(url = list(), ...)
}
\arguments{
\item{url}{See the Stash Playground for details.}

\item{...}{Additional options such as `.field`, `.response`, and `.progress_bar`.}
}
\value{
The processed API response.
}
\description{
Scrapes a complete movie record based on a URL
}
\details{
**Deprecated:** Use scrapeGroupURL instead
}
