% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stashapi_functions.R
\name{scrapePerformerURL}
\alias{scrapePerformerURL}
\title{Call GraphQL operation: scrapePerformerURL}
\usage{
scrapePerformerURL(url = list(), ...)
}
\arguments{
\item{url}{See the Stash Playground for details.}

\item{...}{Additional options such as `.field`, `.response`, and `.progress_bar`.}
}
\value{
The processed API response.
}
\description{
Scrapes a complete performer record based on a URL
}
