% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stashapi_functions.R
\name{scrapeSingleScene}
\alias{scrapeSingleScene}
\title{Call GraphQL operation: scrapeSingleScene}
\usage{
scrapeSingleScene(source = NA, input = NA, ...)
}
\arguments{
\item{source}{See the Playground for further details.}

\item{input}{See the Playground for further details.}

\item{...}{additional parameters like .field to retrieve only a certain field from the response}
}
\value{
processed API response
}
\description{
Scrape for a single scene
}
