% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stashapi_functions.R
\name{scrapeURL}
\alias{scrapeURL}
\title{Call GraphQL operation: scrapeURL}
\usage{
scrapeURL(url = list(), ty = NA, ...)
}
\arguments{
\item{url}{See the Playground for further details.}

\item{ty}{See the Playground for further details.}

\item{...}{additional parameters like .field to retrieve only a certain field from the response}
}
\value{
processed API response
}
\description{
Scrapes content based on a URL
}
