% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stashapi_functions.R
\name{setDefaultFilter}
\alias{setDefaultFilter}
\title{Call GraphQL operation: setDefaultFilter}
\usage{
setDefaultFilter(input = NA, ...)
}
\arguments{
\item{input}{See the Stash Playground for details.}

\item{...}{Additional options such as `.field`, `.response`, and `.progress_bar`.}
}
\value{
The processed API response.
}
\description{
Executes the GraphQL operation `setDefaultFilter`.
}
\details{
**Deprecated:** now uses UI config
}
