% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stashapi_functions.R
\name{setPluginsEnabled}
\alias{setPluginsEnabled}
\title{Call GraphQL operation: setPluginsEnabled}
\usage{
setPluginsEnabled(enabledMap = list(), ...)
}
\arguments{
\item{enabledMap}{See the Stash Playground for details.}

\item{...}{Additional options such as `.field`, `.response`, and `.progress_bar`.}
}
\value{
The processed API response.
}
\description{
Enable/disable plugins - enabledMap is a map of plugin IDs to enabled booleans.
Plugins not in the map are not affected.
}
