% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setStashCredentials.R
\name{stash_disconnect}
\alias{stash_disconnect}
\title{Disconnect the active Stash connection}
\usage{
stash_disconnect()
}
\value{
Invisibly, `TRUE`.
}
\description{
Disconnect the active Stash connection
}
