% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stashapi_functions.R
\name{stashBoxBatchStudioTag}
\alias{stashBoxBatchStudioTag}
\title{Call GraphQL operation: stashBoxBatchStudioTag}
\usage{
stashBoxBatchStudioTag(input = NA, ...)
}
\arguments{
\item{input}{See the Playground for further details.}

\item{...}{additional parameters like .field to retrieve only a certain field from the response}
}
\value{
processed API response
}
\description{
Run batch studio tag task. Returns the job ID.
}
