% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stashapi_functions.R
\name{stopJob}
\alias{stopJob}
\title{Call GraphQL operation: stopJob}
\usage{
stopJob(jobid = list(), ...)
}
\arguments{
\item{jobid}{See the Playground for further details.}

\item{...}{additional parameters like .field to retrieve only a certain field from the response}
}
\value{
processed API response
}
\description{
Call GraphQL operation: stopJob
}
