% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stashapi_functions.R
\name{studiosDestroy}
\alias{studiosDestroy}
\title{Call GraphQL operation: studiosDestroy}
\usage{
studiosDestroy(ids = list(), ...)
}
\arguments{
\item{ids}{See the Playground for further details.}

\item{...}{additional parameters like .field to retrieve only a certain field from the response}
}
\value{
processed API response
}
\description{
Call GraphQL operation: studiosDestroy
}
