% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stashapi_functions.R
\name{studioUpdate}
\alias{studioUpdate}
\title{Call GraphQL operation: studioUpdate}
\usage{
studioUpdate(input = NA, ...)
}
\arguments{
\item{input}{See the Playground for further details.}

\item{...}{additional parameters like .field to retrieve only a certain field from the response}
}
\value{
processed API response
}
\description{
Call GraphQL operation: studioUpdate
}
