% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stashapi_functions.R
\name{submitStashBoxSceneDraft}
\alias{submitStashBoxSceneDraft}
\title{Call GraphQL operation: submitStashBoxSceneDraft}
\usage{
submitStashBoxSceneDraft(input = NA, ...)
}
\arguments{
\item{input}{See the Stash Playground for details.}

\item{...}{Additional options such as `.field`, `.response`, and `.progress_bar`.}
}
\value{
The processed API response.
}
\description{
Submit scene as draft to stash-box instance
}
