% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stashapi_functions.R
\name{addTempDLNAIP}
\alias{addTempDLNAIP}
\title{Call GraphQL operation: addTempDLNAIP}
\usage{
addTempDLNAIP(input = NA, ...)
}
\arguments{
\item{input}{See the Stash Playground for details.}

\item{...}{Additional options such as `.field`, `.response`, and `.progress_bar`.}
}
\value{
The processed API response.
}
\description{
Enables an IP address for DLNA for an optional duration
}
