% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stashapi_functions.R
\name{availablePackages}
\alias{availablePackages}
\title{Call GraphQL operation: availablePackages}
\usage{
availablePackages(type = NA, source = list(), ...)
}
\arguments{
\item{type}{See the Playground for further details.}

\item{source}{See the Playground for further details.}

\item{...}{additional parameters like .field to retrieve only a certain field from the response}
}
\value{
processed API response
}
\description{
List available packages
}
