% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stashapi_functions.R
\name{backupDatabase}
\alias{backupDatabase}
\title{Call GraphQL operation: backupDatabase}
\usage{
backupDatabase(input = NA, ...)
}
\arguments{
\item{input}{See the Stash Playground for details.}

\item{...}{Additional options such as `.field`, `.response`, and `.progress_bar`.}
}
\value{
The processed API response.
}
\description{
Backup the database. Optionally returns a link to download the database file
}
