% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_helpers.R
\name{between}
\alias{between}
\title{Build a BETWEEN criterion.}
\usage{
between(value, value2)
}
\arguments{
\item{value}{Lower or first criterion value.}
\item{value2}{Upper or second criterion value.}
}
\description{
Build a BETWEEN criterion.
}
\examples{
between(1, 10)
}
