% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stashapi_functions.R
\name{bulkSceneMarkerUpdate}
\alias{bulkSceneMarkerUpdate}
\title{Call GraphQL operation: bulkSceneMarkerUpdate}
\usage{
bulkSceneMarkerUpdate(input = NA, ...)
}
\arguments{
\item{input}{See the Playground for further details.}

\item{...}{additional parameters like .field to retrieve only a certain field from the response}
}
\value{
processed API response
}
\description{
Call GraphQL operation: bulkSceneMarkerUpdate
}
