% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stashapi_functions.R
\name{configuration}
\alias{configuration}
\title{Call GraphQL operation: configuration}
\usage{
configuration(...)
}
\arguments{
\item{...}{additional parameters like .field to retrieve only a certain field from the response}
}
\value{
processed API response
}
\description{
Returns the current, complete configuration
}
