% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stashapi_functions.R
\name{configureGeneral}
\alias{configureGeneral}
\title{Call GraphQL operation: configureGeneral}
\usage{
configureGeneral(input = NA, ...)
}
\arguments{
\item{input}{See the Playground for further details.}

\item{...}{additional parameters like .field to retrieve only a certain field from the response}
}
\value{
processed API response
}
\description{
Change general configuration options
}
