% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stashapi_functions.R
\name{configurePlugin}
\alias{configurePlugin}
\title{Call GraphQL operation: configurePlugin}
\usage{
configurePlugin(pluginid = list(), input = list(), ...)
}
\arguments{
\item{pluginid}{See the Stash Playground for details.}

\item{input}{See the Stash Playground for details.}

\item{...}{Additional options such as `.field`, `.response`, and `.progress_bar`.}
}
\value{
The processed API response.
}
\description{
overwrites the entire plugin configuration for the given plugin
}
