% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stashapi_functions.R
\name{configureUI}
\alias{configureUI}
\title{Call GraphQL operation: configureUI}
\usage{
configureUI(input = NA, partial = NA, ...)
}
\arguments{
\item{input}{See the Playground for further details.}

\item{partial}{See the Playground for further details.}

\item{...}{additional parameters like .field to retrieve only a certain field from the response}
}
\value{
processed API response
}
\description{
overwrites the UI configuration
if input is provided, then the entire UI configuration is replaced
if partial is provided, then the partial UI configuration is merged into the existing UI configuration
}
