% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stashapi_functions.R
\name{configureUISetting}
\alias{configureUISetting}
\title{Call GraphQL operation: configureUISetting}
\usage{
configureUISetting(key = list(), value = NA, ...)
}
\arguments{
\item{key}{See the Playground for further details.}

\item{value}{See the Playground for further details.}

\item{...}{additional parameters like .field to retrieve only a certain field from the response}
}
\value{
processed API response
}
\description{
sets a single UI key value
key is a dot separated path to the value
}
