% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stashapi_functions.R
\name{deleteFiles}
\alias{deleteFiles}
\title{Call GraphQL operation: deleteFiles}
\usage{
deleteFiles(ids = list(), ...)
}
\arguments{
\item{ids}{See the Stash Playground for details.}

\item{...}{Additional options such as `.field`, `.response`, and `.progress_bar`.}
}
\value{
The processed API response.
}
\description{
Executes the GraphQL operation `deleteFiles`.
}
