% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stashapi_functions.R
\name{destroyFiles}
\alias{destroyFiles}
\title{Call GraphQL operation: destroyFiles}
\usage{
destroyFiles(ids = list(), ...)
}
\arguments{
\item{ids}{See the Stash Playground for details.}

\item{...}{Additional options such as `.field`, `.response`, and `.progress_bar`.}
}
\value{
The processed API response.
}
\description{
Deletes file entries from the database without deleting the files from the filesystem
}
