% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stashapi_functions.R
\name{destroySavedFilter}
\alias{destroySavedFilter}
\title{Call GraphQL operation: destroySavedFilter}
\usage{
destroySavedFilter(input = NA, ...)
}
\arguments{
\item{input}{See the Playground for further details.}

\item{...}{additional parameters like .field to retrieve only a certain field from the response}
}
\value{
processed API response
}
\description{
Call GraphQL operation: destroySavedFilter
}
