% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stashapi_functions.R
\name{dlnaStatus}
\alias{dlnaStatus}
\title{Call GraphQL operation: dlnaStatus}
\usage{
dlnaStatus(...)
}
\arguments{
\item{...}{Additional options such as `.field`, `.response`, and `.progress_bar`.}
}
\value{
The processed API response.
}
\description{
Executes the GraphQL operation `dlnaStatus`.
}
