% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stashapi_functions.R
\name{downloadFFMpeg}
\alias{downloadFFMpeg}
\title{Call GraphQL operation: downloadFFMpeg}
\usage{
downloadFFMpeg(...)
}
\arguments{
\item{...}{Additional options such as `.field`, `.response`, and `.progress_bar`.}
}
\value{
The processed API response.
}
\description{
Downloads and installs ffmpeg and ffprobe binaries into the configuration directory. Returns the job ID.
}
