% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stashapi_functions.R
\name{exportObjects}
\alias{exportObjects}
\title{Call GraphQL operation: exportObjects}
\usage{
exportObjects(input = NA, ...)
}
\arguments{
\item{input}{See the Stash Playground for details.}

\item{...}{Additional options such as `.field`, `.response`, and `.progress_bar`.}
}
\value{
The processed API response.
}
\description{
Returns a link to download the result
}
