% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stashapi_functions.R
\name{findDuplicateScenes}
\alias{findDuplicateScenes}
\title{Call GraphQL operation: findDuplicateScenes}
\usage{
findDuplicateScenes(distance = NA, durationdiff = NA, ...)
}
\arguments{
\item{distance}{See the Stash Playground for details.}

\item{durationdiff}{Max difference in seconds between files in order to be considered for similarity matching.
Fractional seconds are ok: 0.5 will mean only files that have durations within 0.5 seconds between them will be matched based on PHash distance.}

\item{...}{Additional options such as `.field`, `.response`, and `.progress_bar`.}
}
\value{
The processed API response.
}
\description{
Returns any groups of scenes that are perceptual duplicates within the queried distance
and the difference between their duration is smaller than durationDiff
}
