% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stashapi_functions.R
\name{findDuplicateScenes}
\alias{findDuplicateScenes}
\title{Call GraphQL operation: findDuplicateScenes}
\usage{
findDuplicateScenes(distance = NA, durationdiff = NA, ...)
}
\arguments{
\item{distance}{See the Playground for further details.}

\item{durationdiff}{Max difference in seconds between files in order to be considered for similarity matching.
Fractional seconds are ok: 0.5 will mean only files that have durations within 0.5 seconds between them will be matched based on PHash distance.}

\item{...}{additional parameters including \code{.field} for field selection, \code{.response} for response mode, and \code{.progress_bar} to show an opt-in progress bar while collecting paginated results.}
}
\value{
processed API response
}
\description{
Returns any groups of scenes that are perceptual duplicates within the queried distance
and the difference between their duration is smaller than durationDiff
}
