% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stashapi_functions.R
\name{findFiles}
\alias{findFiles}
\title{Call GraphQL operation: findFiles}
\usage{
findFiles(filefilter = NA, filter = NA, ids = list(), ...)
}
\arguments{
\item{filefilter}{See the Playground for further details.}

\item{filter}{See the Playground for further details.}

\item{ids}{See the Playground for further details.}

\item{...}{additional parameters like .field to retrieve only a certain field from the response}
}
\value{
processed API response
}
\description{
Queries for Files
}
