% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stashapi_functions.R
\name{findFolders}
\alias{findFolders}
\title{Call GraphQL operation: findFolders}
\usage{
findFolders(folderfilter = NA, filter = NA, ids = list(), ...)
}
\arguments{
\item{folderfilter}{See the Playground for further details.}

\item{filter}{See the Playground for further details.}

\item{ids}{See the Playground for further details.}

\item{...}{additional parameters like .field to retrieve only a certain field from the response}
}
\value{
processed API response
}
\description{
Queries for Files
}
