% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stashapi_functions.R
\name{findGroups}
\alias{findGroups}
\title{Call GraphQL operation: findGroups}
\usage{
findGroups(groupfilter = NA, filter = NA, ids = list(), ...)
}
\arguments{
\item{groupfilter}{See the Stash Playground for details.}

\item{filter}{See the Stash Playground for details.}

\item{ids}{See the Stash Playground for details.}

\item{...}{Additional options such as `.field`, `.response`, and `.progress_bar`.}
}
\value{
The processed API response.
}
\description{
A function which queries Group objects
}
