% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stashapi_functions.R
\name{findImage}
\alias{findImage}
\title{Call GraphQL operation: findImage}
\usage{
findImage(id = NA, checksum = NA, ...)
}
\arguments{
\item{id}{See the Playground for further details.}

\item{checksum}{See the Playground for further details.}

\item{...}{additional parameters including \code{.field} for field selection, \code{.response} for response mode, and \code{.progress_bar} to show an opt-in progress bar while collecting paginated results.}
}
\value{
processed API response
}
\description{
Call GraphQL operation: findImage
}
