% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stashapi_functions.R
\name{findImages}
\alias{findImages}
\title{Call GraphQL operation: findImages}
\usage{
findImages(imagefilter = NA, imageids = list(), ids = list(), filter = NA, ...)
}
\arguments{
\item{imagefilter}{See the Stash Playground for details.}

\item{imageids}{See the Stash Playground for details.}

\item{ids}{See the Stash Playground for details.}

\item{filter}{See the Stash Playground for details.}

\item{...}{Additional options such as `.field`, `.response`, and `.progress_bar`.}
}
\value{
The processed API response.
}
\description{
A function which queries Scene objects
}
