% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stashapi_functions.R
\name{findPerformer}
\alias{findPerformer}
\title{Call GraphQL operation: findPerformer}
\usage{
findPerformer(id = list(), ...)
}
\arguments{
\item{id}{See the Playground for further details.}

\item{...}{additional parameters like .field to retrieve only a certain field from the response}
}
\value{
processed API response
}
\description{
Find a performer by ID
}
