% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stashapi_functions.R
\name{findPerformers}
\alias{findPerformers}
\title{Call GraphQL operation: findPerformers}
\usage{
findPerformers(
  performerfilter = NA,
  filter = NA,
  performerids = list(),
  ids = list(),
  ...
)
}
\arguments{
\item{performerfilter}{See the Playground for further details.}

\item{filter}{See the Playground for further details.}

\item{performerids}{See the Playground for further details.}

\item{ids}{See the Playground for further details.}

\item{...}{additional parameters like .field to retrieve only a certain field from the response}
}
\value{
processed API response
}
\description{
A function which queries Performer objects
}
