% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stashapi_functions.R
\name{findPerformers}
\alias{findPerformers}
\title{Call GraphQL operation: findPerformers}
\usage{
findPerformers(
  performerfilter = NA,
  filter = NA,
  performerids = list(),
  ids = list(),
  ...
)
}
\arguments{
\item{performerfilter}{See the Stash Playground for details.}

\item{filter}{See the Stash Playground for details.}

\item{performerids}{See the Stash Playground for details.}

\item{ids}{See the Stash Playground for details.}

\item{...}{Additional options such as `.field`, `.response`, and `.progress_bar`.}
}
\value{
The processed API response.
}
\description{
A function which queries Performer objects
}
