% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stashapi_functions.R
\name{findSavedFilter}
\alias{findSavedFilter}
\title{Call GraphQL operation: findSavedFilter}
\usage{
findSavedFilter(id = list(), ...)
}
\arguments{
\item{id}{See the Playground for further details.}

\item{...}{additional parameters including \code{.field} for field selection, \code{.response} for response mode, and \code{.progress_bar} to show an opt-in progress bar while collecting paginated results.}
}
\value{
processed API response
}
\description{
Call GraphQL operation: findSavedFilter
}
