% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stashapi_functions.R
\name{findSavedFilters}
\alias{findSavedFilters}
\title{Call GraphQL operation: findSavedFilters}
\usage{
findSavedFilters(mode = NA, ...)
}
\arguments{
\item{mode}{See the Stash Playground for details.}

\item{...}{Additional options such as `.field`, `.response`, and `.progress_bar`.}
}
\value{
The processed API response.
}
\description{
Executes the GraphQL operation `findSavedFilters`.
}
