% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stashapi_functions.R
\name{findSavedFilters}
\alias{findSavedFilters}
\title{Call GraphQL operation: findSavedFilters}
\usage{
findSavedFilters(mode = NA, ...)
}
\arguments{
\item{mode}{See the Playground for further details.}

\item{...}{additional parameters like .field to retrieve only a certain field from the response}
}
\value{
processed API response
}
\description{
Call GraphQL operation: findSavedFilters
}
