% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stashapi_functions.R
\name{findScene}
\alias{findScene}
\title{Call GraphQL operation: findScene}
\usage{
findScene(id = NA, checksum = NA, ...)
}
\arguments{
\item{id}{See the Playground for further details.}

\item{checksum}{See the Playground for further details.}

\item{...}{additional parameters like .field to retrieve only a certain field from the response}
}
\value{
processed API response
}
\description{
Find a scene by ID or Checksum
}
