% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stashapi_functions.R
\name{findSceneMarkers}
\alias{findSceneMarkers}
\title{Call GraphQL operation: findSceneMarkers}
\usage{
findSceneMarkers(scenemarkerfilter = NA, filter = NA, ids = list(), ...)
}
\arguments{
\item{scenemarkerfilter}{See the Playground for further details.}

\item{filter}{See the Playground for further details.}

\item{ids}{See the Playground for further details.}

\item{...}{additional parameters like .field to retrieve only a certain field from the response}
}
\value{
processed API response
}
\description{
A function which queries SceneMarker objects
}
