% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stashapi_functions.R
\name{findScenesByPathRegex}
\alias{findScenesByPathRegex}
\title{Call GraphQL operation: findScenesByPathRegex}
\usage{
findScenesByPathRegex(filter = NA, ...)
}
\arguments{
\item{filter}{See the Stash Playground for details.}

\item{...}{Additional options such as `.field`, `.response`, and `.progress_bar`.}
}
\value{
The processed API response.
}
\description{
Executes the GraphQL operation `findScenesByPathRegex`.
}
