% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stashapi_functions.R
\name{findScenesByPathRegex}
\alias{findScenesByPathRegex}
\title{Call GraphQL operation: findScenesByPathRegex}
\usage{
findScenesByPathRegex(filter = NA, ...)
}
\arguments{
\item{filter}{See the Playground for further details.}

\item{...}{additional parameters including \code{.field} for field selection, \code{.response} for response mode, and \code{.progress_bar} to show an opt-in progress bar while collecting paginated results.}
}
\value{
processed API response
}
\description{
Call GraphQL operation: findScenesByPathRegex
}
