% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stashapi_functions.R
\name{findStudios}
\alias{findStudios}
\title{Call GraphQL operation: findStudios}
\usage{
findStudios(studiofilter = NA, filter = NA, ids = list(), ...)
}
\arguments{
\item{studiofilter}{See the Playground for further details.}

\item{filter}{See the Playground for further details.}

\item{ids}{See the Playground for further details.}

\item{...}{additional parameters like .field to retrieve only a certain field from the response}
}
\value{
processed API response
}
\description{
A function which queries Studio objects
}
