% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stashapi_functions.R
\name{findTags}
\alias{findTags}
\title{Call GraphQL operation: findTags}
\usage{
findTags(tagfilter = NA, filter = NA, ids = list(), ...)
}
\arguments{
\item{tagfilter}{See the Stash Playground for details.}

\item{filter}{See the Stash Playground for details.}

\item{ids}{See the Stash Playground for details.}

\item{...}{Additional options such as `.field`, `.response`, and `.progress_bar`.}
}
\value{
The processed API response.
}
\description{
Executes the GraphQL operation `findTags`.
}
