% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stashapi_functions.R
\name{galleryChapterDestroy}
\alias{galleryChapterDestroy}
\title{Call GraphQL operation: galleryChapterDestroy}
\usage{
galleryChapterDestroy(id = list(), ...)
}
\arguments{
\item{id}{See the Stash Playground for details.}

\item{...}{Additional options such as `.field`, `.response`, and `.progress_bar`.}
}
\value{
The processed API response.
}
\description{
Executes the GraphQL operation `galleryChapterDestroy`.
}
