% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_helpers.R
\name{greater_than}
\alias{greater_than}
\title{Build a greater-than criterion.}
\usage{
greater_than(value)
}
\arguments{
\item{value}{Criterion value.}
}
\description{
The helper name matches the GraphQL schema modifier.
}
\examples{
greater_than("FULL_HD")
}
