% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_helpers.R
\name{group_filter}
\alias{group_filter}
\title{Build a group filter.}
\usage{group_filter(..., .strict = TRUE)}
\arguments{\item{...}{Filter fields from the Stash `GroupFilterType` input.}\item{.strict}{Whether unknown fields should cause an error.}}
\value{A named list suitable for `findGroups(groupfilter = ...)`.}
