% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setStashCredentials.R
\name{hasConnection}
\alias{hasConnection}
\title{Check if a Valid Connection to Stash Exists}
\usage{
hasConnection()
}
\value{
A logical value:
- `TRUE` if a valid connection exists
- `FALSE` if no connection is found or connection validation fails
}
\description{
Validates whether a connection to the GraphQL endpoint of Stash can be established and is functional.
}
\examples{
\dontrun{
	hasConnection()
}

}
\keyword{connection}
\keyword{validation}
