% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_helpers.R
\name{image_filter}
\alias{image_filter}
\title{Build an image filter.}
\usage{image_filter(..., .strict = TRUE)}
\arguments{
\item{...}{Filter fields from the Stash `ImageFilterType` input.}
\item{.strict}{Whether unknown fields should cause an error.}
}
\value{A named list suitable for `findImages(imagefilter = ...)`.}
