% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stashapi_functions.R
\name{imageDecrementO}
\alias{imageDecrementO}
\title{Call GraphQL operation: imageDecrementO}
\usage{
imageDecrementO(id = list(), ...)
}
\arguments{
\item{id}{See the Playground for further details.}

\item{...}{additional parameters like .field to retrieve only a certain field from the response}
}
\value{
processed API response
}
\description{
Decrements the o-counter for an image. Returns the new value
}
