% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stashapi_functions.R
\name{imageIncrementO}
\alias{imageIncrementO}
\title{Call GraphQL operation: imageIncrementO}
\usage{
imageIncrementO(id = list(), ...)
}
\arguments{
\item{id}{See the Playground for further details.}

\item{...}{additional parameters like .field to retrieve only a certain field from the response}
}
\value{
processed API response
}
\description{
Increments the o-counter for an image. Returns the new value
}
