% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stashapi_functions.R
\name{importObjects}
\alias{importObjects}
\title{Call GraphQL operation: importObjects}
\usage{
importObjects(input = NA, ...)
}
\arguments{
\item{input}{See the Stash Playground for details.}

\item{...}{Additional options such as `.field`, `.response`, and `.progress_bar`.}
}
\value{
The processed API response.
}
\description{
Performs an incremental import. Returns the job ID
}
