% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_helpers.R
\name{includes}
\alias{includes}
\title{Build an INCLUDES criterion.}
\usage{
includes(value, depth = NULL, excludes = NULL)
}
\arguments{
\item{value}{Criterion value.}

\item{depth}{Optional hierarchy depth. `-1` means unlimited depth.}

\item{excludes}{Optional IDs to exclude.}
}
\description{
Build an INCLUDES criterion.
}
\examples{
includes(182, depth = -1)
}
