% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stashapi_functions.R
\name{installedPackages}
\alias{installedPackages}
\title{Call GraphQL operation: installedPackages}
\usage{
installedPackages(type = NA, ...)
}
\arguments{
\item{type}{See the Playground for further details.}

\item{...}{additional parameters like .field to retrieve only a certain field from the response}
}
\value{
processed API response
}
\description{
List installed packages
}
