% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stashapi_functions.R
\name{installPackages}
\alias{installPackages}
\title{Call GraphQL operation: installPackages}
\usage{
installPackages(type = NA, packages = NA, ...)
}
\arguments{
\item{type}{See the Stash Playground for details.}

\item{packages}{See the Stash Playground for details.}

\item{...}{Additional options such as `.field`, `.response`, and `.progress_bar`.}
}
\value{
The processed API response.
}
\description{
Installs the given packages.
If a package is already installed, it will be updated if needed..
If an error occurs when installing a package, the job will continue to install the remaining packages.
Returns the job ID
}
