% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_helpers.R
\name{is_null}
\alias{is_null}
\title{Build an IS_NULL criterion.}
\usage{
is_null(value = TRUE)
}
\arguments{
\item{value}{Boolean null criterion value.}
}
\description{
Build an IS_NULL criterion.
}
\examples{
is_null()
}
