% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stashapi_functions.R
\name{latestversion}
\alias{latestversion}
\title{Call GraphQL operation: latestversion}
\usage{
latestversion(...)
}
\arguments{
\item{...}{additional parameters like .field to retrieve only a certain field from the response}
}
\value{
processed API response
}
\description{
Call GraphQL operation: latestversion
}
