% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_helpers.R
\name{less_than}
\alias{less_than}
\title{Build a less-than criterion.}
\usage{
less_than(value)
}
\arguments{\item{value}{Criterion value.}}
\description{
Build a less-than criterion.
}
\examples{
less_than(720)
}
