% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stashapi_functions.R
\name{listScrapers}
\alias{listScrapers}
\title{Call GraphQL operation: listScrapers}
\usage{
listScrapers(types = NA, ...)
}
\arguments{
\item{types}{See the Playground for further details.}

\item{...}{additional parameters like .field to retrieve only a certain field from the response}
}
\value{
processed API response
}
\description{
List available scrapers
}
