% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stashapi_functions.R
\name{markerStrings}
\alias{markerStrings}
\title{Call GraphQL operation: markerStrings}
\usage{
markerStrings(q = NA, sort = NA, ...)
}
\arguments{
\item{q}{See the Playground for further details.}

\item{sort}{See the Playground for further details.}

\item{...}{additional parameters like .field to retrieve only a certain field from the response}
}
\value{
processed API response
}
\description{
Get marker strings
}
