% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stashapi_functions.R
\name{markerWall}
\alias{markerWall}
\title{Call GraphQL operation: markerWall}
\usage{
markerWall(q = NA, ...)
}
\arguments{
\item{q}{See the Playground for further details.}

\item{...}{additional parameters like .field to retrieve only a certain field from the response}
}
\value{
processed API response
}
\description{
Retrieve random scene markers for the wall
}
