% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_helpers.R
\name{matches_regex}
\alias{matches_regex}
\title{Build a regular-expression match criterion.}
\usage{
matches_regex(value)
}
\arguments{\item{value}{Regular-expression pattern.}}
\description{
Build a regular-expression match criterion.
}
\examples{
matches_regex("^4k")
}
