% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stashapi_functions.R
\name{metadataAutoTag}
\alias{metadataAutoTag}
\title{Call GraphQL operation: metadataAutoTag}
\usage{
metadataAutoTag(input = NA, ...)
}
\arguments{
\item{input}{See the Stash Playground for details.}

\item{...}{Additional options such as `.field`, `.response`, and `.progress_bar`.}
}
\value{
The processed API response.
}
\description{
Start auto-tagging. Returns the job ID
}
