% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stashapi_functions.R
\name{metadataGenerate}
\alias{metadataGenerate}
\title{Call GraphQL operation: metadataGenerate}
\usage{
metadataGenerate(input = NA, ...)
}
\arguments{
\item{input}{See the Stash Playground for details.}

\item{...}{Additional options such as `.field`, `.response`, and `.progress_bar`.}
}
\value{
The processed API response.
}
\description{
Start generating content. Returns the job ID
}
