% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stashapi_functions.R
\name{metadataImport}
\alias{metadataImport}
\title{Call GraphQL operation: metadataImport}
\usage{
metadataImport(...)
}
\arguments{
\item{...}{Additional options such as `.field`, `.response`, and `.progress_bar`.}
}
\value{
The processed API response.
}
\description{
Start an full import. Completely wipes the database and imports from the metadata directory. Returns the job ID
}
