% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stashapi_functions.R
\name{metadataScan}
\alias{metadataScan}
\title{Call GraphQL operation: metadataScan}
\usage{
metadataScan(input = NA, ...)
}
\arguments{
\item{input}{See the Playground for further details.}

\item{...}{additional parameters like .field to retrieve only a certain field from the response}
}
\value{
processed API response
}
\description{
Start a scan. Returns the job ID
}
