% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stashapi_functions.R
\name{migrate}
\alias{migrate}
\title{Call GraphQL operation: migrate}
\usage{
migrate(input = NA, ...)
}
\arguments{
\item{input}{See the Stash Playground for details.}

\item{...}{Additional options such as `.field`, `.response`, and `.progress_bar`.}
}
\value{
The processed API response.
}
\description{
Migrates the schema to the required version. Returns the job ID
}
