% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stashapi_functions.R
\name{migrateBlobs}
\alias{migrateBlobs}
\title{Call GraphQL operation: migrateBlobs}
\usage{
migrateBlobs(input = NA, ...)
}
\arguments{
\item{input}{See the Playground for further details.}

\item{...}{additional parameters like .field to retrieve only a certain field from the response}
}
\value{
processed API response
}
\description{
Migrates blobs from the old storage system to the current one
}
