% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stashapi_functions.R
\name{migrateHashNaming}
\alias{migrateHashNaming}
\title{Call GraphQL operation: migrateHashNaming}
\usage{
migrateHashNaming(...)
}
\arguments{
\item{...}{additional parameters like .field to retrieve only a certain field from the response}
}
\value{
processed API response
}
\description{
Migrate generated files for the current hash naming
}
