% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stashapi_functions.R
\name{moveFiles}
\alias{moveFiles}
\title{Call GraphQL operation: moveFiles}
\usage{
moveFiles(input = NA, ...)
}
\arguments{
\item{input}{See the Playground for further details.}

\item{...}{additional parameters like .field to retrieve only a certain field from the response}
}
\value{
processed API response
}
\description{
Moves the given files to the given destination. Returns true if successful.
Either the destination_folder or destination_folder_id must be provided.
If both are provided, the destination_folder_id takes precedence.
Destination folder must be a subfolder of one of the stash library paths.
If provided, destination_basename must be a valid filename with an extension that
matches one of the media extensions.
Creates folder hierarchy if needed.
}
