% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_helpers.R
\name{not_equals}
\alias{not_equals}
\title{Build a NOT_EQUALS criterion.}
\usage{
not_equals(value)
}
\arguments{\item{value}{Criterion value.}}
\value{A named criterion list.}
\description{Builds a GraphQL NOT_EQUALS criterion.}
