% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_helpers.R
\name{not_matches_regex}
\alias{not_matches_regex}
\title{Build a negated regular-expression match criterion.}
\usage{
not_matches_regex(value)
}
\arguments{\item{value}{Regular-expression pattern.}}
\description{
Build a negated regular-expression match criterion.
}
\examples{
not_matches_regex("sample")
}
