% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_helpers.R
\name{not_null}
\alias{not_null}
\title{Build a NOT_NULL criterion.}
\usage{
not_null(value = NULL)
}
\arguments{\item{value}{Optional criterion value.}}
\description{
Build a NOT_NULL criterion.
}
\examples{not_null()}
