% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stashapi_functions.R
\name{optimiseDatabase}
\alias{optimiseDatabase}
\title{Call GraphQL operation: optimiseDatabase}
\usage{
optimiseDatabase(...)
}
\arguments{
\item{...}{additional parameters like .field to retrieve only a certain field from the response}
}
\value{
processed API response
}
\description{
Optimises the database. Returns the job ID
}
