% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stashapi_functions.R
\name{parseSceneFilenames}
\alias{parseSceneFilenames}
\title{Call GraphQL operation: parseSceneFilenames}
\usage{
parseSceneFilenames(filter = NA, config = NA, ...)
}
\arguments{
\item{filter}{See the Playground for further details.}

\item{config}{See the Playground for further details.}

\item{...}{additional parameters like .field to retrieve only a certain field from the response}
}
\value{
processed API response
}
\description{
Call GraphQL operation: parseSceneFilenames
}
