% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stashapi_functions.R
\name{plugins}
\alias{plugins}
\title{Call GraphQL operation: plugins}
\usage{
plugins(...)
}
\arguments{
\item{...}{additional parameters like .field to retrieve only a certain field from the response}
}
\value{
processed API response
}
\description{
List loaded plugins
}
