% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stashapi_functions.R
\name{pluginTasks}
\alias{pluginTasks}
\title{Call GraphQL operation: pluginTasks}
\usage{
pluginTasks(...)
}
\arguments{
\item{...}{additional parameters like .field to retrieve only a certain field from the response}
}
\value{
processed API response
}
\description{
List available plugin operations
}
