% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stashapi_functions.R
\name{querySQL}
\alias{querySQL}
\title{Call GraphQL operation: querySQL}
\usage{
querySQL(sql = list(), args = NA, ...)
}
\arguments{
\item{sql}{See the Stash Playground for details.}

\item{args}{See the Stash Playground for details.}

\item{...}{Additional options such as `.field`, `.response`, and `.progress_bar`.}
}
\value{
The processed API response.
}
\description{
DANGEROUS: Execute an arbitrary SQL statement that returns rows.
}
