% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stashapi_functions.R
\name{removeTempDLNAIP}
\alias{removeTempDLNAIP}
\title{Call GraphQL operation: removeTempDLNAIP}
\usage{
removeTempDLNAIP(input = NA, ...)
}
\arguments{
\item{input}{See the Playground for further details.}

\item{...}{additional parameters like .field to retrieve only a certain field from the response}
}
\value{
processed API response
}
\description{
Removes an IP address from the temporary DLNA whitelist
}
