% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stashapi_functions.R
\name{reorderSubGroups}
\alias{reorderSubGroups}
\title{Call GraphQL operation: reorderSubGroups}
\usage{
reorderSubGroups(input = NA, ...)
}
\arguments{
\item{input}{See the Playground for further details.}

\item{...}{additional parameters like .field to retrieve only a certain field from the response}
}
\value{
processed API response
}
\description{
Reorder sub groups within a group. Returns true if successful.
}
