% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stashapi_functions.R
\name{revealFileInFileManager}
\alias{revealFileInFileManager}
\title{Call GraphQL operation: revealFileInFileManager}
\usage{
revealFileInFileManager(id = list(), ...)
}
\arguments{
\item{id}{See the Stash Playground for details.}

\item{...}{Additional options such as `.field`, `.response`, and `.progress_bar`.}
}
\value{
The processed API response.
}
\description{
Reveal the file in the system file manager
}
