% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stashapi_functions.R
\name{runPluginOperation}
\alias{runPluginOperation}
\title{Call GraphQL operation: runPluginOperation}
\usage{
runPluginOperation(pluginid = list(), args = NA, ...)
}
\arguments{
\item{pluginid}{See the Stash Playground for details.}

\item{args}{See the Stash Playground for details.}

\item{...}{Additional options such as `.field`, `.response`, and `.progress_bar`.}
}
\value{
The processed API response.
}
\description{
Runs a plugin operation. The operation is run immediately and does not use the job queue.
Returns a map of the result.
}
