% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stashapi_functions.R
\name{saveFilter}
\alias{saveFilter}
\title{Call GraphQL operation: saveFilter}
\usage{
saveFilter(input = NA, ...)
}
\arguments{
\item{input}{See the Stash Playground for details.}

\item{...}{Additional options such as `.field`, `.response`, and `.progress_bar`.}
}
\value{
The processed API response.
}
\description{
Executes the GraphQL operation `saveFilter`.
}
