% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stashapi_functions.R
\name{sceneAddO}
\alias{sceneAddO}
\title{Call GraphQL operation: sceneAddO}
\usage{
sceneAddO(id = list(), times = list(), ...)
}
\arguments{
\item{id}{See the Stash Playground for details.}

\item{times}{See the Stash Playground for details.}

\item{...}{Additional options such as `.field`, `.response`, and `.progress_bar`.}
}
\value{
The processed API response.
}
\description{
Increments the o-counter for a scene. Uses the current time if none provided.
}
