% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stashapi_functions.R
\name{sceneAddPlay}
\alias{sceneAddPlay}
\title{Call GraphQL operation: sceneAddPlay}
\usage{
sceneAddPlay(id = list(), times = list(), ...)
}
\arguments{
\item{id}{See the Playground for further details.}

\item{times}{See the Playground for further details.}

\item{...}{additional parameters like .field to retrieve only a certain field from the response}
}
\value{
processed API response
}
\description{
Increments the play count for the scene. Uses the current time if none provided.
}
