% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stashapi_functions.R
\name{sceneDeleteO}
\alias{sceneDeleteO}
\title{Call GraphQL operation: sceneDeleteO}
\usage{
sceneDeleteO(id = list(), times = list(), ...)
}
\arguments{
\item{id}{See the Stash Playground for details.}

\item{times}{See the Stash Playground for details.}

\item{...}{Additional options such as `.field`, `.response`, and `.progress_bar`.}
}
\value{
The processed API response.
}
\description{
Decrements the o-counter for a scene, removing the last recorded time if specific time not provided. Returns the new value
}
