% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stashapi_functions.R
\name{sceneDeletePlay}
\alias{sceneDeletePlay}
\title{Call GraphQL operation: sceneDeletePlay}
\usage{
sceneDeletePlay(id = list(), times = list(), ...)
}
\arguments{
\item{id}{See the Stash Playground for details.}

\item{times}{See the Stash Playground for details.}

\item{...}{Additional options such as `.field`, `.response`, and `.progress_bar`.}
}
\value{
The processed API response.
}
\description{
Decrements the play count for the scene, removing the specific times or the last recorded time if not provided.
}
