% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stashapi_functions.R
\name{sceneDestroy}
\alias{sceneDestroy}
\title{Call GraphQL operation: sceneDestroy}
\usage{
sceneDestroy(input = NA, ...)
}
\arguments{
\item{input}{See the Playground for further details.}

\item{...}{additional parameters like .field to retrieve only a certain field from the response}
}
\value{
processed API response
}
\description{
Call GraphQL operation: sceneDestroy
}
