% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stashapi_functions.R
\name{sceneGenerateScreenshot}
\alias{sceneGenerateScreenshot}
\title{Call GraphQL operation: sceneGenerateScreenshot}
\usage{
sceneGenerateScreenshot(id = list(), at = NA, ...)
}
\arguments{
\item{id}{See the Stash Playground for details.}

\item{at}{See the Stash Playground for details.}

\item{...}{Additional options such as `.field`, `.response`, and `.progress_bar`.}
}
\value{
The processed API response.
}
\description{
Generates screenshot at specified time in seconds. Leave empty to generate default screenshot
}
