% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stashapi_functions.R
\name{sceneMarkerDestroy}
\alias{sceneMarkerDestroy}
\title{Call GraphQL operation: sceneMarkerDestroy}
\usage{
sceneMarkerDestroy(id = list(), ...)
}
\arguments{
\item{id}{See the Stash Playground for details.}

\item{...}{Additional options such as `.field`, `.response`, and `.progress_bar`.}
}
\value{
The processed API response.
}
\description{
Executes the GraphQL operation `sceneMarkerDestroy`.
}
