% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stashapi_functions.R
\name{sceneMarkersDestroy}
\alias{sceneMarkersDestroy}
\title{Call GraphQL operation: sceneMarkersDestroy}
\usage{
sceneMarkersDestroy(ids = list(), ...)
}
\arguments{
\item{ids}{See the Playground for further details.}

\item{...}{additional parameters like .field to retrieve only a certain field from the response}
}
\value{
processed API response
}
\description{
Call GraphQL operation: sceneMarkersDestroy
}
