% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stashapi_functions.R
\name{sceneMarkerTags}
\alias{sceneMarkerTags}
\title{Call GraphQL operation: sceneMarkerTags}
\usage{
sceneMarkerTags(sceneid = list(), ...)
}
\arguments{
\item{sceneid}{See the Playground for further details.}

\item{...}{additional parameters like .field to retrieve only a certain field from the response}
}
\value{
processed API response
}
\description{
Organize scene markers by tag for a given scene ID
}
