% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stashapi_functions.R
\name{sceneResetActivity}
\alias{sceneResetActivity}
\title{Call GraphQL operation: sceneResetActivity}
\usage{
sceneResetActivity(id = list(), resetresume = NA, resetduration = NA, ...)
}
\arguments{
\item{id}{See the Playground for further details.}

\item{resetresume}{See the Playground for further details.}

\item{resetduration}{See the Playground for further details.}

\item{...}{additional parameters like .field to retrieve only a certain field from the response}
}
\value{
processed API response
}
\description{
Resets the resume time point and play duration
}
