% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stashapi_functions.R
\name{sceneResetO}
\alias{sceneResetO}
\title{Call GraphQL operation: sceneResetO}
\usage{
sceneResetO(id = list(), ...)
}
\arguments{
\item{id}{See the Playground for further details.}

\item{...}{additional parameters like .field to retrieve only a certain field from the response}
}
\value{
processed API response
}
\description{
Resets the o-counter for a scene to 0. Returns the new value
}
