% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stashapi_functions.R
\name{sceneSaveActivity}
\alias{sceneSaveActivity}
\title{Call GraphQL operation: sceneSaveActivity}
\usage{
sceneSaveActivity(id = list(), resumetime = NA, playDuration = NA, ...)
}
\arguments{
\item{id}{See the Stash Playground for details.}

\item{resumetime}{See the Stash Playground for details.}

\item{playDuration}{See the Stash Playground for details.}

\item{...}{Additional options such as `.field`, `.response`, and `.progress_bar`.}
}
\value{
The processed API response.
}
\description{
Sets the resume time point (if provided) and adds the provided duration to the scene's play duration
}
