% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stashapi_functions.R
\name{sceneStreams}
\alias{sceneStreams}
\title{Call GraphQL operation: sceneStreams}
\usage{
sceneStreams(id = NA, ...)
}
\arguments{
\item{id}{See the Playground for further details.}

\item{...}{additional parameters like .field to retrieve only a certain field from the response}
}
\value{
processed API response
}
\description{
Return valid stream paths
}
