% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stashapi_functions.R
\name{sceneWall}
\alias{sceneWall}
\title{Call GraphQL operation: sceneWall}
\usage{
sceneWall(q = NA, ...)
}
\arguments{
\item{q}{See the Stash Playground for details.}

\item{...}{Additional options such as `.field`, `.response`, and `.progress_bar`.}
}
\value{
The processed API response.
}
\description{
Retrieve random scenes for the wall
}
