% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stashapi_functions.R
\name{scrapeMultiPerformers}
\alias{scrapeMultiPerformers}
\title{Call GraphQL operation: scrapeMultiPerformers}
\usage{
scrapeMultiPerformers(source = NA, input = NA, ...)
}
\arguments{
\item{source}{See the Playground for further details.}

\item{input}{See the Playground for further details.}

\item{...}{additional parameters like .field to retrieve only a certain field from the response}
}
\value{
processed API response
}
\description{
Scrape for multiple performers
}
