% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stashapi_functions.R
\name{scrapeMultiScenes}
\alias{scrapeMultiScenes}
\title{Call GraphQL operation: scrapeMultiScenes}
\usage{
scrapeMultiScenes(source = NA, input = NA, ...)
}
\arguments{
\item{source}{See the Playground for further details.}

\item{input}{See the Playground for further details.}

\item{...}{additional parameters like .field to retrieve only a certain field from the response}
}
\value{
processed API response
}
\description{
Scrape for multiple scenes
}
