% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stashapi_functions.R
\name{scrapePerformerURL}
\alias{scrapePerformerURL}
\title{Call GraphQL operation: scrapePerformerURL}
\usage{
scrapePerformerURL(url = list(), ...)
}
\arguments{
\item{url}{See the Playground for further details.}

\item{...}{additional parameters like .field to retrieve only a certain field from the response}
}
\value{
processed API response
}
\description{
Scrapes a complete performer record based on a URL
}
