% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stashapi_functions.R
\name{scrapeSinglePerformer}
\alias{scrapeSinglePerformer}
\title{Call GraphQL operation: scrapeSinglePerformer}
\usage{
scrapeSinglePerformer(source = NA, input = NA, ...)
}
\arguments{
\item{source}{See the Stash Playground for details.}

\item{input}{See the Stash Playground for details.}

\item{...}{Additional options such as `.field`, `.response`, and `.progress_bar`.}
}
\value{
The processed API response.
}
\description{
Scrape for a single performer
}
