% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stashapi_functions.R
\name{scrapeURL}
\alias{scrapeURL}
\title{Call GraphQL operation: scrapeURL}
\usage{
scrapeURL(url = list(), ty = NA, ...)
}
\arguments{
\item{url}{See the Stash Playground for details.}

\item{ty}{See the Stash Playground for details.}

\item{...}{Additional options such as `.field`, `.response`, and `.progress_bar`.}
}
\value{
The processed API response.
}
\description{
Scrapes content based on a URL
}
