% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stashapi_functions.R
\name{setPluginsEnabled}
\alias{setPluginsEnabled}
\title{Call GraphQL operation: setPluginsEnabled}
\usage{
setPluginsEnabled(enabledMap = list(), ...)
}
\arguments{
\item{enabledMap}{See the Playground for further details.}

\item{...}{additional parameters like .field to retrieve only a certain field from the response}
}
\value{
processed API response
}
\description{
Enable/disable plugins - enabledMap is a map of plugin IDs to enabled booleans.
Plugins not in the map are not affected.
}
