% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_helpers.R
\name{stash_id}
\alias{stash_id}
\title{Build a Stash ID criterion.}
\usage{stash_id(endpoint, stash_id, modifier = "EQUALS")}
\arguments{
\item{endpoint}{Stash endpoint URL.}
\item{stash_id}{Endpoint-specific Stash ID.}
\item{modifier}{GraphQL criterion modifier.}
}
\value{A named Stash ID criterion list.}
