% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_helpers.R
\name{stash_ids}
\alias{stash_ids}
\title{Build a multiple Stash ID criterion.}
\usage{stash_ids(endpoint, stash_ids, modifier = "EQUALS")}
\arguments{
\item{endpoint}{Stash endpoint URL.}
\item{stash_ids}{Endpoint-specific Stash IDs.}
\item{modifier}{GraphQL criterion modifier.}
}
\value{A named multiple Stash ID criterion list.}
