% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stashapi_functions.R
\name{stashBoxBatchStudioTag}
\alias{stashBoxBatchStudioTag}
\title{Call GraphQL operation: stashBoxBatchStudioTag}
\usage{
stashBoxBatchStudioTag(input = NA, ...)
}
\arguments{
\item{input}{See the Stash Playground for details.}

\item{...}{Additional options such as `.field`, `.response`, and `.progress_bar`.}
}
\value{
The processed API response.
}
\description{
Run batch studio tag task. Returns the job ID.
}
