% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stashapi_functions.R
\name{stats}
\alias{stats}
\title{Call GraphQL operation: stats}
\usage{
stats(...)
}
\arguments{
\item{...}{Additional options such as `.field`, `.response`, and `.progress_bar`.}
}
\value{
The processed API response.
}
\description{
Get stats
}
