% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stashapi_functions.R
\name{stopJob}
\alias{stopJob}
\title{Call GraphQL operation: stopJob}
\usage{
stopJob(jobid = list(), ...)
}
\arguments{
\item{jobid}{See the Stash Playground for details.}

\item{...}{Additional options such as `.field`, `.response`, and `.progress_bar`.}
}
\value{
The processed API response.
}
\description{
Executes the GraphQL operation `stopJob`.
}
