% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stashapi_functions.R
\name{submitStashBoxPerformerDraft}
\alias{submitStashBoxPerformerDraft}
\title{Call GraphQL operation: submitStashBoxPerformerDraft}
\usage{
submitStashBoxPerformerDraft(input = NA, ...)
}
\arguments{
\item{input}{See the Stash Playground for details.}

\item{...}{Additional options such as `.field`, `.response`, and `.progress_bar`.}
}
\value{
The processed API response.
}
\description{
Submit performer as draft to stash-box instance
}
