% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stashapi_functions.R
\name{submitStashBoxSceneDraft}
\alias{submitStashBoxSceneDraft}
\title{Call GraphQL operation: submitStashBoxSceneDraft}
\usage{
submitStashBoxSceneDraft(input = NA, ...)
}
\arguments{
\item{input}{See the Playground for further details.}

\item{...}{additional parameters like .field to retrieve only a certain field from the response}
}
\value{
processed API response
}
\description{
Submit scene as draft to stash-box instance
}
