% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stashapi_functions.R
\name{systemStatus}
\alias{systemStatus}
\title{Call GraphQL operation: systemStatus}
\usage{
systemStatus(...)
}
\arguments{
\item{...}{Additional options such as `.field`, `.response`, and `.progress_bar`.}
}
\value{
The processed API response.
}
\description{
Executes the GraphQL operation `systemStatus`.
}
