% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_helpers.R
\name{tag_filter}
\alias{tag_filter}
\title{Build a tag filter.}
\usage{
tag_filter(..., .strict = TRUE)
}
\arguments{
\item{...}{Filter fields from the Stash `TagFilterType` input.}

\item{.strict}{Whether unknown fields should cause an error.}
}
\value{
A named list suitable for `findTags(tagfilter = ...)`.
}
\description{
Build a tag filter.
}
