% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stashapi_functions.R
\name{uninstallPackages}
\alias{uninstallPackages}
\title{Call GraphQL operation: uninstallPackages}
\usage{
uninstallPackages(type = NA, packages = NA, ...)
}
\arguments{
\item{type}{See the Stash Playground for details.}

\item{packages}{See the Stash Playground for details.}

\item{...}{Additional options such as `.field`, `.response`, and `.progress_bar`.}
}
\value{
The processed API response.
}
\description{
Uninstalls the given packages.
If an error occurs when uninstalling a package, the job will continue to uninstall the remaining packages.
Returns the job ID
}
