% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stashapi_functions.R
\name{updatePackages}
\alias{updatePackages}
\title{Call GraphQL operation: updatePackages}
\usage{
updatePackages(type = NA, packages = NA, ...)
}
\arguments{
\item{type}{See the Playground for further details.}

\item{packages}{See the Playground for further details.}

\item{...}{additional parameters like .field to retrieve only a certain field from the response}
}
\value{
processed API response
}
\description{
Updates the given packages.
If a package is not installed, it will not be installed.
If a package does not need to be updated, it will not be updated.
If no packages are provided, all packages of the given type will be updated.
If an error occurs when updating a package, the job will continue to update the remaining packages.
Returns the job ID.
}
